code_assistant/requirements.txt
streamlit>=1.34
plotly>=5.22
reportlab>=4.0

pydantic>=2.7
PyYAML>=6.0
tqdm>=4.66

ruff>=0.5
mypy>=1.10
bandit>=1.7
pip-audit>=2.7
radon>=6.0

pytest>=8.2
coverage>=7.6
hypothesis>=6.111
